pyproject.toml
[project]
name = "agents-sdk-research"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.0.3",
    "python-dotenv>=1.0.1",
]
